first commit
This is first working version of my idea. Simply, give all the chain of
select boxes, define URLs for each segment, define default selected
option values for each segments and that’s it. It doesn’t matter two or
more select boxes. As long as you define the URLs, it keep segmenting
to next select box you defined and load content to the end.